requirements.txt
asgiref==3.8.1
Authlib==1.3.1
backports.zoneinfo==0.2.1;python_version<"3.9"
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.0
cryptography==43.0.0
distlib==0.3.8
Django==4.2.14
django-cors-headers==4.4.0
django-allauth==0.63.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
filelock==3.15.4
idna==3.7
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyproject-api==1.7.1
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.1
tox==4.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
